requirements.txt
# IA e Agentes
anthropic>=0.92.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0

# Geração de imagens
openai>=1.50.0
google-genai>=1.0.0

# Google Ads
google-ads>=24.0.0

# Banco de dados
sqlalchemy>=2.0.0

# Utilitários
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
pydantic>=2.0.0

# API
fastapi>=0.115.0
uvicorn>=0.30.0

# Dashboard
streamlit>=1.35.0
plotly>=5.22.0

# Web (landing page agent)
requests>=2.31.0
beautifulsoup4>=4.12.0
